Create NCARBFilter.html
This is the html code for the NCARB Assessment.
The JSON file is embedded in the code. It filters the users' list using  first name, last name, city, state code and country code.
When no parameter is entered all the names (20 in all) are displayed. The ng-repeat returns the number of rows of users that match the entered parameter values